findCategoryForTransaction (with updated Regex)
I've taken each keyword from the transaction description, searched in the keyword repository using findByValueContaining, like searching for "Carrefour" in the list of keywords.

The thing is, for example, it's POSSIBLE that multiple keywords appear in that description, so we're looking for the keyword that is most frequently found among all. That way, we assign that keyword and its category.

Another thing would be that in the end, we CLEAN UP unnecessary keywords like: "srl", "sa", "romania", "sc", "impex".

MAYBE needing updating at TransactionDTO